All works as needed to pass tests

+++ Recipes/task/src/recipes/services/RecipeService.java
@@ -1,9 +1,14 @@
 package recipes.services;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import recipes.model.Recipe;
+import recipes.model.User;
 import recipes.repos.RecipesCrudRepo;
+import recipes.repos.UserRepo;
 
 import java.time.LocalDateTime;
 import java.util.Comparator;
@@ -14,10 +19,12 @@ import java.util.stream.Collectors;
 @Service
 public class RecipeService {
     RecipesCrudRepo recipeRepo;
+    UserRepo userRepo;
 
     @Autowired
-    public RecipeService(RecipesCrudRepo recipeRepo) {
+    public RecipeService(RecipesCrudRepo recipeRepo, UserRepo userRepo ) {
         this.recipeRepo = recipeRepo;
+        this.userRepo = userRepo;
     }
 
     public Optional<Recipe> getRecipe(long id) {
@@ -26,40 +33,51 @@ public class RecipeService {
 
     public long setRecipe(Recipe recipe) {
         recipe.setDate(LocalDateTime.now());
+        recipe.setAuthor(getCurrentUser());
         return recipeRepo.save(recipe).getId();
     }
 
-    public boolean deleteRecipe(long id) {
+    public HttpStatus deleteRecipe(long id) {
         Optional<Recipe> recipeOptional = recipeRepo.findById(id);
         if (recipeOptional.isPresent()) {
-            recipeRepo.delete(recipeOptional.get());
-            return true;
+            User currentUser = getCurrentUser();
+            if (currentUser == recipeOptional.get().getAuthor()) {
+                recipeRepo.delete(recipeOptional.get());
+                return HttpStatus.NO_CONTENT;
+            }
+            return HttpStatus.FORBIDDEN;
         }
-        return false;
+        return HttpStatus.NOT_FOUND;
+    }
+
+    private User getCurrentUser() {
+        Authentication authentication = SecurityContextHolder.getContext()
+                .getAuthentication();
+        return userRepo.findByEmail(authentication.getName()).get();
     }
 
 
-    public boolean updateRecipe(long id, Recipe updatedRecipe) {
+    public HttpStatus updateRecipe(long id, Recipe updatedRecipe) {
         Optional<Recipe> optionalRecipe = recipeRepo.findById(id);
         if (optionalRecipe.isPresent()) {
-            Recipe recipe = optionalRecipe.get();
-            recipe.setDate(LocalDateTime.now());
-            recipe.setCategory(updatedRecipe.getCategory());
-            recipe.setDescription(updatedRecipe.getDescription());
-            recipe.setDirections(updatedRecipe.getDirections());
-            recipe.setIngredients(updatedRecipe.getIngredients());
-            recipe.setName(updatedRecipe.getName());
-            recipeRepo.save(recipe);
-            return true;
+            if (getCurrentUser() == optionalRecipe.get().getAuthor()) {
+                Recipe recipe = optionalRecipe.get();
+                recipe.setDate(LocalDateTime.now());
+                recipe.setCategory(updatedRecipe.getCategory());
+                recipe.setDescription(updatedRecipe.getDescription());
+                recipe.setDirections(updatedRecipe.getDirections());
+                recipe.setIngredients(updatedRecipe.getIngredients());
+                recipe.setName(updatedRecipe.getName());
+                recipe.setAuthor(getCurrentUser());
+                recipeRepo.save(recipe);
+                return HttpStatus.NO_CONTENT;
+            }
+            return HttpStatus.FORBIDDEN;
         }
-        return false;
+        return HttpStatus.NOT_FOUND;
     }
 
     public List<Recipe> findByName(String name) {
-        //var recipes = recipeRepo.findByNameIgnoreCase("\"%" + name + "%\"");
-        //var recipes = recipeRepo.findAll();
-        //recipes.sort(Comparator.comparing(Recipe::getDate));
-        //recipes.forEach(System.out::println);
 
         return recipeRepo.findAll().stream().filter(r -> r.getName().toUpperCase()
                 .matches("(.+ +)?" + name.toUpperCase() + "( +.+)?"))
@@ -69,10 +87,6 @@ public class RecipeService {
     }
 
     public List<Recipe> findByCategory(String category) {
-        //var recipes = recipeRepo.findByCategoryIgnoreCase("\"%" + category + "%\"");
-        //System.out.println("\"%" + category + "%\"");
-        //recipes.sort(Comparator.comparing(Recipe::getDate));
-        //recipes.forEach(System.out::println);
         return  recipeRepo.findAll().stream().filter(r -> r.getCategory().toUpperCase()
                 .matches("(.+ +)?" + category.toUpperCase() + ".*"))
                 .sorted(Comparator.comparing(Recipe::getDate).reversed())
